requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart>=0.0.5,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
langchain>=0.0.300
langchain-community>=0.0.1
langchain-openai>=0.0.1
langgraph>=0.0.1
faiss-cpu>=1.7.4
openai>=1.0.0
langfuse>=2.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0 